pyproject.toml
[tool.poetry]
name = "workshop-b2"
version = "0.1.0"
description = ""
authors = ["OpsMill <info@opsmill.com>"]
readme = "README.md"
packages = [
    { include = "workshop_b2" }
]

[tool.poetry.dependencies]
python = "^3.10, < 3.13"
faker = "^30.8.1"
fastapi = {extras = ["standard"], version = "^0.115.5"}
httpx = "^0.27.2"
sqlmodel = "^0.0.22"
typer = "^0.12.5"
uvicorn = "^0.32.0"
strawberry-graphql = "^0.247.0"
invoke = "^2.2.0"
neo4j = "^5.26.0"
infrahub-sdk = {version = "^1.0.1", extras = ["all"]}


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
pytest = "^8.3.3"
invoke = "^2.2.0"
pyright = "^1.1.386"
yamllint = "^1.35.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
